Rename operator log to logscomponent
