+++ src/components/logdisplay/LogDisplay.vue
@@ -1,12 +1,12 @@
 <template>
-  <OperatorLog v-if="logDisplay" :logDisplay />
+  <LogsComponent v-if="logDisplay" :logDisplay />
   <div v-else>
     <v-alert type="warning" class="ma-3">No logs display available</v-alert>
   </div>
 </template>
 
 <script setup lang="ts">
-import OperatorLog from '@/components/logdisplay/OperatorLog.vue'
+import LogsComponent from '@/components/logdisplay/LogsComponent.vue'
 import { configManager } from '@/services/application-config'
 import { useLogDisplay } from '@/services/useLogDisplay'
 import type { TopologyNode } from '@deltares/fews-pi-requests'
