+++ src/components/logdisplay/LogsComponent.vue
